fix rss
