+++ website/ui/components/button.tsx
@@ -1,4 +1,5 @@
 import { ComponentProps, ReactElement } from 'react';
+import NextLink from 'next/link';
 import clsx from 'clsx';
 import { styled } from '../../stitches.config';
 
@@ -64,3 +65,35 @@ export const Button = ({
     </ButtonWithAsProp>
   );
 };
+
+export function GuildButton({
+  children,
+  className,
+  // @ts-expect-error -- fixme
+  as: Component = NextLink,
+  ...props
+}: ComponentProps<typeof NextLink>): ReactElement {
+  return (
+    // @ts-expect-error -- fixme
+    <Component
+      {...props}
+      className={clsx(
+        'text-dark relative inline-flex items-center gap-3 rounded-[10px] px-5 py-3 font-medium transition-none',
+        // @ts-expect-error -- fixme
+        props.disabled
+          ? 'opacity-50'
+          : [
+              'after:font-mono after:text-3xl after:leading-none after:content-["➔"]',
+              'hocus:after:transition-[opacity]',
+              'lg:hocus:pr-10 duration-700 lg:transition-[padding]',
+              'hocus:after:delay-100 hocus:after:duration-500',
+              'lg:hocus:after:opacity-100 lg:after:opacity-0',
+              'lg:after:absolute lg:after:right-3',
+            ],
+        className,
+      )}
+    >
+      {children}
+    </Component>
+  );
+}
