+++ website/ui/components/index.ts
@@ -1,6 +1,6 @@
 export { Avatar } from './avatar';
 export { BlogCardList } from './blog-card-list';
-export { Button } from './button';
+export { Button, GuildButton } from './button';
 export { Comparison } from './comparison';
 export { CopyToClipboard } from './copy-to-clipboard';
 export { Description } from './description';
